Refactored folder structures; Now user can't access dashboard without signing in
