+++ package.json
@@ -21,6 +21,7 @@
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-firebase-hooks": "^5.1.1",
     "react-router-dom": "^6.14.1",
     "sort-by": "^1.2.0"
   },
