+++ src/components/Layout/Layout.jsx
@@ -1,7 +1,6 @@
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { Outlet } from "react-router-dom";
 import Footer from "./components/Footer";
-
 const defaultTheme = createTheme();
 
 function App() {
