+++ src/components/Layout/components/Footer.jsx
@@ -1,6 +1,6 @@
 import { Container, Typography } from "@mui/material";
 import { Box } from "@mui/system";
-import Copyright from "./Copyright";
+import Copyright from "../../Copyright";
 
 function Footer() {
   return (
