--- src/pages/Dashboard.jsx
@@ -1,17 +1,0 @@
-import { useEffect } from "react";
-import { auth } from "../backend/db";
-import SignOutBtn from "../components/SignOutBtn";
-
-function Dashboard() {
-  useEffect(() => {
-    console.log(auth.email, auth);
-  });
-  return (
-    <div>
-      Hello user: {auth}
-      <SignOutBtn />
-    </div>
-  );
-}
-
-export default Dashboard;
