+++ src/pages/Error/Error.jsx
@@ -1,6 +1,6 @@
 import { useRouteError } from "react-router-dom";
 
-export default function ErrorPage() {
+export default function Error() {
   const error = useRouteError();
   console.error(error);
