+++ src/pages/SignUp/SignUp.jsx
@@ -12,7 +12,7 @@ import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
 import { Link as RouterLink } from "react-router-dom";
 import { createUserWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../backend/db";
+import { auth } from "../../backend/db";
 import { useNavigate } from "react-router-dom";
 
 export default function SignUp() {
